Rename nv_memory_options.h to options_nv_memory.h for MPLAB-X projects .
