+++ MatrixPilot/MatrixPilot-udb4.X/nbproject/configurations.xml
@@ -21,7 +21,6 @@
         </logicalFolder>
         <itemPath>../../Config/FSconfig.h</itemPath>
         <itemPath>../../Config/mavlink_options.h</itemPath>
-        <itemPath>../../Config/nv_memory_options.h</itemPath>
         <itemPath>../../Config/options.h</itemPath>
         <itemPath>../../Config/options_auav3.h</itemPath>
         <itemPath>../../Config/waypoints.h</itemPath>
@@ -31,6 +30,7 @@
         <itemPath>../../Config/options_servo_mix.h</itemPath>
         <itemPath>../../Config/options_osd.h</itemPath>
         <itemPath>../../Config/options_ports.h</itemPath>
+        <itemPath>../../Config/options_nv_memory.h</itemPath>
       </logicalFolder>
       <logicalFolder name="libDCM" displayName="libDCM" projectFiles="true">
         <itemPath>../../libDCM/dcmTypes.h</itemPath>
